Slightly restructure modules
